IS_package/requirements.txt
pandas==2.2.2
openpyxl==3.1.5
PyYAML==6.0.2
